requirements.txt
python-telegram-bot==20.8
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
loguru==0.7.2 